+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q003/Solution2.java
@@ -1,9 +1,10 @@
-package com.noob.algorithm.leetcode.q03;
+package com.noob.algorithm.leetcode.q003;
 
 import java.util.HashSet;
 
 /**
- * 003.无重复字符的最大子串
+ * 🟡 003 无重复最长子串
+ * https://leetcode.cn/problems/longest-substring-without-repeating-characters/
  * 思路：滑动窗口
  * 使用一个HashSet来实现滑动窗口，用来检查重复字符。 维护开始和结束两个索引，默认都是从0开始，然后随着循环【向右移动结束索引】，遇到不是重复字符则放入窗里，遇到重复字符则【向右侧移动开始索引】，最终得到结果
  */
